+++ src/test/resources/org/sonar/plugins/delphi/symbol/unitAliases/Unit3.pas
@@ -1,0 +1,21 @@
+unit Unit3;
+
+{This is a sample Delphi file.}
+
+interface
+
+procedure Unit3Proc(Argument: String);
+
+implementation
+
+uses
+    Unit1
+  , Unit2
+  ;
+
+procedure Unit3Proc(Argument: String);
+begin
+  // Do nothing
+end;
+
+end.