+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for unit aliases.
+- `sonar.delphi.unitAliases` property to specify unit aliases, which are used for unit import
+  resolution.
+
 ## [0.12.1] - 2020-02-18
 
 ### Fixed